multisig pallet config refactor
