+++ ownership-chain/runtime/laos/src/configs/mod.rs
@@ -1,1 +1,2 @@
 mod proxy;
+mod multisig;
